+++ packages/elementary/lib/elementary.dart
@@ -2,6 +2,7 @@ library elementary;
 
 export 'package:elementary/src/core.dart' hide Elementary;
 export 'package:elementary/src/error/error_handler.dart';
+export 'package:elementary/src/keep_alive/automatic_keep_alive_widget_model_mixin.dart';
 export 'package:elementary/src/relations/builder/state_notifier_builder.dart';
 export 'package:elementary/src/relations/notifier/state_notifier.dart';
 export 'package:elementary/src/ticker/ticker_providers.dart';